part2/requirements.txt
django==2.2.9
django-rest-framework
# filters support for django
django-filter